feat: refactor account and session entities, remove user entity and related fields
